requirements.txt
# Core dependencies
chainlit>=1.2.0
openai>=1.30.0
python-dotenv>=1.0.0

# Database (SQLite is built into Python)
# No additional dependencies needed for SQLite

# Optional: GCP dependencies for deployment (only needed for production)
google-cloud-storage>=2.10.0
google-cloud-secret-manager>=2.16.4

# Utilities
httpx>=0.28.0